backend/requirements.txt
aliyun-python-sdk-core==2.13.19
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
aniso8601==8.0.0
appnope==0.1.0
art==4.7
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
chrhyme==0.2.6
click==7.1.2
crcmod==1.7
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Flask==0.11.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-WTF==0.14.3
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mongoengine==0.20.0
numpy==1.18.5
OpenCC==1.1.1
opencv-python==4.2.0.34
oss2==2.12.1
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pinyin==0.4.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycryptodome==3.9.8
PyExecJS==1.5.1
Pygments==2.6.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pypinyin==0.38.0
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
structlog==20.1.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
zope.event==4.4
zope.interface==5.1.0
